+++ examples/basic/src/model-provider/openai/openai-chat-stream-text-n-2-example.ts
@@ -8,7 +8,7 @@ async function main() {
     openai.ChatTextGenerator({
       model: "gpt-3.5-turbo",
       maxCompletionTokens: 1000,
-      n: 2,
+      numberOfGenerations: 2,
     }),
     [
       OpenAIChatMessage.system("You are a story writer. Write a story about:"),