+++ src/model-provider/openai/chat/AbstractOpenAIChatModel.ts
@@ -42,19 +42,7 @@ export interface AbstractOpenAIChatCallSettings {
     | "none"
     | "auto"
     | { type: "function"; function: { name: string } };
-  /**
-   * An array of strings or a single string that the model will recognize as end-of-text indicators.
-   * The model stops generating more content when it encounters any of these strings.
-   * This is particularly useful in scripted or formatted text generation, where a specific end point is required.
-   * Example: stop: ['\n', 'END']
-   */
-  stop?: string | string[];
-  /**
-   * Specifies the maximum number of tokens (words, punctuation, parts of words) that the model can generate in a single response.
-   * It helps to control the length of the output, this can help prevent wasted time and tokens when tweaker topP or temperature.
-   * Example: maxTokens: 1000
-   */
-  maxTokens?: number;
+
   /**
    * `temperature`: Controls the randomness and creativity in the model's responses.
    * A lower temperature (close to 0) results in more predictable, conservative text, while a higher temperature (close to 1) produces more varied and creative output.
@@ -62,6 +50,7 @@ export interface AbstractOpenAIChatCallSettings {
    * Example: temperature: 0.5
    */
   temperature?: number;
+
   /**
    *  This parameter sets a threshold for token selection based on probability.
    * The model will only consider the most likely tokens that cumulatively exceed this threshold while generating a response.
@@ -70,6 +59,7 @@ export interface AbstractOpenAIChatCallSettings {
    * Example: topP: 0.2
    */
   topP?: number;
+
   /**
    * Used to set the initial state for the random number generator in the model.
    * Providing a specific seed value ensures consistent outputs for the same inputs across different runs - useful for testing and reproducibility.
@@ -77,18 +67,7 @@ export interface AbstractOpenAIChatCallSettings {
    * Example: seed: 89 (or) seed: null
    */
   seed?: number | null;
-  responseFormat?: {
-    type?: "text" | "json_object";
-  };
 
-  /**
-   * Specifies the number of responses or completions the model should generate for a given prompt.
-   * This is useful when you need multiple different outputs or ideas for a single prompt.
-   * The model will generate 'n' distinct responses, each based on the same initial prompt.
-   * In a streaming model this will result in both responses streamed back in real time.
-   * Example: n: 3 // The model will produce 3 different responses.
-   */
-  n?: number;
   /**
    * Discourages the model from repeating the same information or context already mentioned in the conversation or prompt.
    * Increasing this value encourages the model to introduce new topics or ideas, rather than reiterating what has been said.
@@ -96,6 +75,7 @@ export interface AbstractOpenAIChatCallSettings {
    * Example: presencePenalty: 1.0 // Strongly discourages repeating the same content.
    */
   presencePenalty?: number;
+
   /**
    * This parameter reduces the likelihood of the model repeatedly using the same words or phrases in its responses.
    * A higher frequency penalty promotes a wider variety of language and expressions in the output.
@@ -103,12 +83,17 @@ export interface AbstractOpenAIChatCallSettings {
    * Example: frequencyPenalty: 0.5 // Moderately discourages repetitive language.
    */
   frequencyPenalty?: number;
+
+  responseFormat?: {
+    type?: "text" | "json_object";
+  };
+
   logitBias?: Record<number, number>;
 }
 
 export interface AbstractOpenAIChatSettings
   extends TextGenerationModelSettings,
-    Omit<AbstractOpenAIChatCallSettings, "stop" | "maxTokens"> {
+    AbstractOpenAIChatCallSettings {
   isUserIdForwardingEnabled?: boolean;
 }
 
@@ -137,37 +122,67 @@ export abstract class AbstractOpenAIChatModel<
         toolChoice?: AbstractOpenAIChatCallSettings["toolChoice"];
       }
   ): Promise<RESULT> {
+    const api = this.settings.api ?? new OpenAIApiConfiguration();
+    const responseFormat = options.responseFormat;
+    const abortSignal = options.run?.abortSignal;
+    const user = this.settings.isUserIdForwardingEnabled
+      ? options.run?.userId
+      : undefined;
+    const openAIResponseFormat = this.settings.responseFormat;
+
+    // function & tool calling:
+    const functions = options.functions ?? this.settings.functions;
+    const functionCall = options.functionCall ?? this.settings.functionCall;
+    const tools = options.tools ?? this.settings.tools;
+    const toolChoice = options.toolChoice ?? this.settings.toolChoice;
+
+    let { stopSequences } = this.settings;
+
     return callWithRetryAndThrottle({
       retry: this.settings.api?.retry,
       throttle: this.settings.api?.throttle,
-      call: async () =>
-        callOpenAIChatCompletionAPI({
-          ...this.settings,
+      call: async () => {
+        // empty arrays are not allowed for stopSequences:
+        if (
+          stopSequences != null &&
+          Array.isArray(stopSequences) &&
+          stopSequences.length === 0
+        ) {
+          stopSequences = undefined;
+        }
 
-          // function & tool calling:
-          functions: options.functions ?? this.settings.functions,
-          functionCall: options.functionCall ?? this.settings.functionCall,
-          tools: options.tools ?? this.settings.tools,
-          toolChoice: options.toolChoice ?? this.settings.toolChoice,
-
-          // map to OpenAI API names:
-          stop: this.settings.stopSequences,
-          maxTokens: this.settings.maxCompletionTokens,
-          openAIResponseFormat: this.settings.responseFormat,
-
-          // other settings:
-          user: this.settings.isUserIdForwardingEnabled
-            ? options.run?.userId
-            : undefined,
-          abortSignal: options.run?.abortSignal,
-
-          responseFormat: options.responseFormat,
-          messages,
-        }),
+        return postJsonToApi({
+          url: api.assembleUrl("/chat/completions"),
+          headers: api.headers,
+          body: {
+            stream: responseFormat.stream,
+            model: this.settings.model,
+            messages,
+            functions,
+            function_call: functionCall,
+            tools,
+            tool_choice: toolChoice,
+            temperature: this.settings.temperature,
+            top_p: this.settings.topP,
+            n: this.settings.numberOfGenerations,
+            stop: this.settings.stopSequences,
+            max_tokens: this.settings.maxCompletionTokens,
+            presence_penalty: this.settings.presencePenalty,
+            frequency_penalty: this.settings.frequencyPenalty,
+            logit_bias: this.settings.logitBias,
+            seed: this.settings.seed,
+            response_format: openAIResponseFormat,
+            user,
+          },
+          failedResponseHandler: failedOpenAICallResponseHandler,
+          successfulResponseHandler: responseFormat.handler,
+          abortSignal,
+        });
+      },
     });
   }
 
-  async doGenerateText(prompt: OpenAIChatPrompt, options?: FunctionOptions) {
+  async doGenerateTexts(prompt: OpenAIChatPrompt, options?: FunctionOptions) {
     const response = await this.callAPI(prompt, {
       ...options,
       responseFormat: OpenAIChatResponseFormat.json,
@@ -175,7 +190,7 @@ export abstract class AbstractOpenAIChatModel<
 
     return {
       response,
-      text: response.choices[0]!.message.content!,
+      texts: response.choices.map((choice) => choice.message.content ?? ""),
       usage: this.extractUsage(response),
     };
   }
@@ -321,69 +336,6 @@ const openAIChatResponseSchema = z.object({
 });
 
 export type OpenAIChatResponse = z.infer<typeof openAIChatResponseSchema>;
-
-async function callOpenAIChatCompletionAPI<RESPONSE>({
-  api = new OpenAIApiConfiguration(),
-  abortSignal,
-  responseFormat,
-  model,
-  messages,
-  functions,
-  functionCall,
-  tools,
-  toolChoice,
-  temperature,
-  topP,
-  n,
-  stop,
-  maxTokens,
-  presencePenalty,
-  frequencyPenalty,
-  logitBias,
-  user,
-  openAIResponseFormat,
-  seed,
-}: AbstractOpenAIChatCallSettings & {
-  api?: ApiConfiguration;
-  abortSignal?: AbortSignal;
-  responseFormat: OpenAIChatResponseFormatType<RESPONSE>;
-  messages: Array<OpenAIChatMessage>;
-  user?: string;
-  openAIResponseFormat: AbstractOpenAIChatCallSettings["responseFormat"]; // mapping
-}): Promise<RESPONSE> {
-  // empty arrays are not allowed for stop:
-  if (stop != null && Array.isArray(stop) && stop.length === 0) {
-    stop = undefined;
-  }
-
-  return postJsonToApi({
-    url: api.assembleUrl("/chat/completions"),
-    headers: api.headers,
-    body: {
-      stream: responseFormat.stream,
-      model,
-      messages,
-      functions,
-      function_call: functionCall,
-      tools,
-      tool_choice: toolChoice,
-      temperature,
-      top_p: topP,
-      n,
-      stop,
-      max_tokens: maxTokens,
-      presence_penalty: presencePenalty,
-      frequency_penalty: frequencyPenalty,
-      logit_bias: logitBias,
-      seed,
-      response_format: openAIResponseFormat,
-      user,
-    },
-    failedResponseHandler: failedOpenAICallResponseHandler,
-    successfulResponseHandler: responseFormat.handler,
-    abortSignal,
-  });
-}
 
 export type OpenAIChatResponseFormatType<T> = {
   stream: boolean;